Disable filling the local cache in getRoute
The getRoute function was filling the local cache from KVs (Key Value
Store) when the cache was initially empty. As the number of entries in
the KV grew, this became very high on the CPU.

This commit removes this logic.

It's worth noticing that cloudflare worker processes seem quite short
lived, so a large pertcentage of those processes weren't used more than
once. As such, loading the whole content of the KV database is really the
opposite of an optimization.